myApp/requirements.txt
django
djangorestframework
mysqlclient
# psycopg2